refactor: Extract shared Social Cards UI from design studio
Refs: #324
