+++ explorer/app/streamlit/design_share_summary_app.py
@@ -23,20 +23,29 @@ if _REPO_ROOT not in sys.path:
 import pandas as pd
 import streamlit as st
 
+from explorer.app.streamlit.social_cards_session_keys import DESIGN_SOCIAL_CARDS_KEYS
+from explorer.app.streamlit.social_cards_sidebar_ui import (
+    render_sidebar_card_controls,
+    render_sidebar_geo_scope_controls,
+)
+from explorer.app.streamlit.social_cards_streamlit_helpers import (
+    card_stat_data_scope_from_design_source,
+    period_has_checklist_data,
+)
+from explorer.app.streamlit.social_cards_streamlit_ui import (
+    SOCIAL_CARDS_CURRENT_CARD_LABEL,
+    render_current_card_fragment,
+)
 from explorer.core.data_loader import add_datetime_column, load_dataset
-from explorer.core.region_display import map_focus_key_for_display
 from explorer.core.settings_schema_defaults import TAXONOMY_LOCALE_DEFAULT
 from explorer.core.share_summary_compute import (
     PeriodAnchor,
-    PeriodKind,
     ShareSummaryAllTimeStats,
     ShareSummaryGeoScope,
     ShareSummaryStats,
     compute_share_summary_all_time_stats,
     compute_share_summary_stats,
     filter_df_by_geo_scope,
-    geo_country_keys_from_df,
-    geo_region_options_for_country,
     geo_scope_display_label,
     period_for_custom,
     period_for_lifetime,
@@ -47,32 +56,18 @@ from explorer.core.share_summary_compute import (
     suggest_period_anchor,
 )
 from explorer.core.share_summary_defaults import (
-    SHARE_SUMMARY_COLOR_SCHEME_IDS,
     SHARE_SUMMARY_INSIGHT_FACT_DEFAULT,
     SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT,
     SHARE_SUMMARY_STORY_MAX_STATS,
-    share_summary_color_scheme_fingerprint,
-    share_summary_color_scheme_index,
-    share_summary_color_scheme_label,
-    share_summary_layout_label,
-    share_summary_spotlight_presentation_label,
-    share_summary_tiles_presentation_label,
 )
 from explorer.core.share_summary_insight_facts import (
-    INSIGHT_FACT_PICKER_LABELS,
-    InsightFactId,
     ShareSummaryInsightFact,
     compute_insight_facts,
-    insight_fact_requires_species,
     species_common_names_in_period,
 )
 from explorer.presentation.share_summary_circle_layout_playground import (
     CIRCLE_LAYOUT_PLAYGROUND_IFRAME_HEIGHT_PX,
     render_circle_layout_playground_html,
-)
-from explorer.presentation.share_summary_circles_preview import (
-    TILES_CIRCLE_CLUSTER_DEFAULT,
-    tiles_circle_cluster_max,
 )
 from explorer.presentation.share_summary_hex_preview import (
     HEX_GRID_EXPERIMENTAL_NOTICE,
@@ -81,84 +76,19 @@ from explorer.presentation.share_summary_hex_preview import (
     HexVariantId,
     render_hex_grid_preview_html,
 )
-from explorer.presentation.share_summary_png_export import (
-    share_summary_png_filename,
-    share_summary_to_png_bytes,
-)
 from explorer.presentation.share_summary_preview import (
-    FORMAT_LABELS,
-    FORMAT_PIXELS,
-    FormatId,
-    LayoutId,
-    SpotlightPresentationId,
-    TilesPresentationId,
     default_card_stat_labels,
-    layout_card_stat_max,
-    layout_card_stat_storage_max,
-    layout_grid_slot_limits_caption,
-    layout_grid_stat_default_count,
-    layout_grid_stat_min,
-    render_share_summary_preview_html,
-    resolve_insight_fact,
     sample_share_summary_stats,
     summary_status_metrics,
 )
-
-
-@st.cache_data(show_spinner="Generating PNG…")
-def _cached_share_summary_png(
-    stats: ShareSummaryStats,
-    layout: LayoutId,
-    fmt: FormatId,
-    card_stat_labels: tuple[str, ...],
-    spotlight_label: str,
-    all_time: ShareSummaryAllTimeStats | None,
-    color_scheme_index: int,
-    color_scheme_fingerprint: tuple[tuple[str, str], ...],
-    scope_label: str | None,
-    geo_scope: ShareSummaryGeoScope,
-    tiles_presentation: TilesPresentationId = "grid",
-    spotlight_presentation: SpotlightPresentationId = "classic",
-    insight_fact: ShareSummaryInsightFact | None = None,
-) -> bytes:
-    del color_scheme_fingerprint  # cache key only — render reads live scheme by index
-    return share_summary_to_png_bytes(
-        stats,
-        layout=layout,
-        fmt=fmt,
-        spotlight_label=spotlight_label,
-        insight_fact=insight_fact,
-        card_stat_labels=card_stat_labels,
-        all_time=all_time,
-        color_scheme_index=color_scheme_index,
-        scope_label=scope_label,
-        geo_scope=geo_scope,
-        tiles_presentation=tiles_presentation,
-        spotlight_presentation=spotlight_presentation,
-    )
 
 _DESIGN_STUDIO_TITLE = "Social sharing design studio"
 _SOCIAL_CARDS_TAB_LABEL = "Social Cards"
 _HEX_EXPERIMENTS_TAB_LABEL = "Hex grid experiments (experimental)"
 _CIRCLE_LAYOUT_TAB_LABEL = "Circle layout"
-_PREVIEW_SCALE_DEFAULT = 0.42
-_PREVIEW_SCALE_FULL = 1.0
-_TILES_PRESENTATION_KEY = "design_tiles_presentation"
-_SPOTLIGHT_PRESENTATION_KEY = "design_spotlight_presentation"
-_INSIGHT_FACT_KEY = "design_insight_fact"
-_INSIGHT_SPECIES_KEY = "design_insight_species"
-_COLOR_THEME_KEY = "design_color_theme"
-_STATISTICS_LABEL = "Card statistics"
-_CARD_STATS_SLOT_COUNT_PREFIX = "design_card_stat_slot_count_"
-_CARD_STATS_PICKS_PREFIX = "design_card_stat_picks_"
-_CARD_STATS_SCOPE_PREFIX = "design_card_stat_scope_"
-_SPOTLIGHT_LABEL_KEY = "design_spotlight_label"
-_CURRENT_CARD_LABEL = "Current card"
 _CARD_HEADING_LABEL = "Card Heading (optional)"
 _CARD_HEADING_PLACEHOLDER = "e.g. North Coast NSW Exploration"
-_GEO_WORLD_OPTION = ""
-_GEO_COUNTRY_SELECT_KEY = "design_geo_country"
-_GEO_REGION_SELECT_KEY = "design_geo_region"
+_DESIGN_HEX_SELECTED_KEY = "design_hex_selected_variant"
 
 
 def _card_heading_or_none(text: str) -> str | None:
@@ -250,1008 +180,13 @@ def _design_sample_dataset(sample_year: int) -> pd.DataFrame:
     return add_datetime_column(pd.DataFrame(records))
 
 
-def _geo_country_select_label(country_key: str) -> str:
-    if not country_key:
-        return "World"
-    return map_focus_key_for_display(country_key)
-
-
-def _sidebar_geo_scope_controls(df: pd.DataFrame) -> ShareSummaryGeoScope:
-    """Country and region pickers; World is the default (no filter)."""
-    country_keys = geo_country_keys_from_df(df)
-    country_options = [_GEO_WORLD_OPTION, *country_keys]
-    country_key = st.selectbox(
-        "Country",
-        options=country_options,
-        format_func=_geo_country_select_label,
-        key=_GEO_COUNTRY_SELECT_KEY,
-    )
-    if not country_key:
-        st.session_state.pop(_GEO_REGION_SELECT_KEY, None)
-        return ShareSummaryGeoScope()
-
-    region_pairs = geo_region_options_for_country(df, country_key)
-    region_options = [_GEO_WORLD_OPTION, *[code for code, _ in region_pairs]]
-    region_labels = {_GEO_WORLD_OPTION: "All regions"}
-    region_labels.update({code: label for code, label in region_pairs})
-    current_region = st.session_state.get(_GEO_REGION_SELECT_KEY, _GEO_WORLD_OPTION)
-    if current_region not in region_options:
-        st.session_state[_GEO_REGION_SELECT_KEY] = _GEO_WORLD_OPTION
-    region_code = st.selectbox(
-        "Region",
-        options=region_options,
-        format_func=lambda code: region_labels[code],
-        key=_GEO_REGION_SELECT_KEY,
-    )
-    return ShareSummaryGeoScope(
-        country_key=country_key,
-        region_code=region_code or None,
-    )
-
-
-def _card_stat_picks_key(layout: LayoutId, period_kind: PeriodKind) -> str:
-    return f"{_CARD_STATS_PICKS_PREFIX}{layout}_{period_kind}"
-
-
-def _card_stat_slot_count_key(layout: LayoutId, period_kind: PeriodKind) -> str:
-    return f"{_CARD_STATS_SLOT_COUNT_PREFIX}{layout}_{period_kind}"
-
-
-def _card_stat_scope_key(layout: LayoutId, period_kind: PeriodKind) -> str:
-    return f"{_CARD_STATS_SCOPE_PREFIX}{layout}_{period_kind}"
-
-
-def _card_stat_data_scope(
-    *,
-    use_sample: bool,
-    period_kind: PeriodKind,
-    period_label: str,
-    upload_name: str | None,
-    geo_scope: ShareSummaryGeoScope | None = None,
-    fmt: FormatId = "square",
-    tiles_presentation: TilesPresentationId = "grid",
-) -> str:
-    """Session scope token — when this changes, card-stat picks re-initialize."""
-    source = "sample" if use_sample else (upload_name or "csv")
-    geo_token = (geo_scope or ShareSummaryGeoScope()).scope_token()
-    return f"{source}|{period_kind}|{period_label}|{geo_token}|{fmt}|{tiles_presentation}"
-
-
-def _period_has_checklist_data(stats: ShareSummaryStats) -> bool:
-    """False when the selected period has no checklists in the loaded export."""
-    return stats.checklists is not None and stats.checklists > 0
-
-
-def _card_stat_selectbox_key(layout: LayoutId, index: int) -> str:
-    return f"design_card_stat_sel_{layout}_{index}"
-
-
-def _clear_card_stat_selectbox_keys(layout: LayoutId) -> None:
-    """Drop stale selectbox widget state so Reset / session picks take effect."""
-    for i in range(layout_card_stat_storage_max(layout)):
-        st.session_state.pop(_card_stat_selectbox_key(layout, i), None)
-
-
-def _resolve_card_stat_selectbox_value(
-    *,
-    session_value: object,
-    desired: str,
-    options: list[str],
-) -> str:
-    """Pick a valid selectbox value, preferring the widget over stale session picks."""
-    if session_value is not None:
-        value = session_value.strip() if isinstance(session_value, str) else str(session_value).strip()
-        if value in options:
-            return value
-    fallback = desired.strip() if isinstance(desired, str) else str(desired).strip()
-    return fallback if fallback in options else ""
-
-
-def _sync_card_stat_selectbox_value(
-    layout: LayoutId,
-    index: int,
-    *,
-    options: list[str],
-    desired: str,
-) -> str:
-    """Align selectbox session state without clobbering a valid user choice."""
-    key = _card_stat_selectbox_key(layout, index)
-    value = _resolve_card_stat_selectbox_value(
-        session_value=st.session_state.get(key),
-        desired=desired,
-        options=options,
-    )
-    st.session_state[key] = value
-    return value
-
-
-def _tiles_circle_cluster_picker(layout: LayoutId, tiles_presentation: TilesPresentationId) -> bool:
-    return layout == "tiles" and tiles_presentation == "circles"
-
-
-# Statistics Grid vs circle-cluster slot limits (see layout_grid_* in preview).
-
-
-def _card_stat_min_slots(
-    layout: LayoutId,
-    fmt: FormatId,
-    *,
-    tiles_presentation: TilesPresentationId = "grid",
-) -> int:
-    if layout == "tiles" and not _tiles_circle_cluster_picker(layout, tiles_presentation):
-        return layout_grid_stat_min(fmt)
-    return 1
-
-
-def _default_card_stat_slot_count(
-    *,
-    circle_cluster: bool,
-    defaults: list[str],
-    max_slots: int,
-    layout: LayoutId,
-    fmt: FormatId,
-    tiles_presentation: TilesPresentationId = "grid",
-) -> int:
-    if circle_cluster:
-        return min(max_slots, TILES_CIRCLE_CLUSTER_DEFAULT)
-    if layout == "tiles" and tiles_presentation == "grid":
-        target = layout_grid_stat_default_count(fmt)
-        return min(max_slots, max(layout_grid_stat_min(fmt), target))
-    return min(max_slots, max(1, len(defaults) if defaults else 1))
-
-
-def _card_stat_max_slots(
-    layout: LayoutId,
-    fmt: FormatId,
-    *,
-    tiles_presentation: TilesPresentationId = "grid",
-    status_metrics: list[tuple[str, str]] | None = None,
-    available_stat_count: int | None = None,
-) -> int:
-    if _tiles_circle_cluster_picker(layout, tiles_presentation):
-        return tiles_circle_cluster_max(fmt)
-    count = available_stat_count
-    if count is None and status_metrics is not None:
-        count = len(status_metrics)
-    return layout_card_stat_max(layout, fmt, available_stat_count=count)
-
-
-def _card_stat_ui_row_count(
-    layout: LayoutId,
-    fmt: FormatId,
-    *,
-    slot_count: int,
-    tiles_presentation: TilesPresentationId = "grid",
-    status_metrics: list[tuple[str, str]] | None = None,
-    available_stat_count: int | None = None,
-) -> int:
-    max_slots = _card_stat_max_slots(
-        layout,
-        fmt,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-        available_stat_count=available_stat_count,
-    )
-    min_slots = _card_stat_min_slots(layout, fmt, tiles_presentation=tiles_presentation)
-    return min(max(min_slots, slot_count), max_slots)
-
-
-def _effective_card_stat_labels(
-    picks: list[str],
-    layout: LayoutId,
-    fmt: FormatId,
-    *,
-    tiles_presentation: TilesPresentationId = "grid",
-    status_metrics: list[tuple[str, str]] | None = None,
-) -> tuple[str, ...]:
-    max_slots = _card_stat_max_slots(
-        layout, fmt, tiles_presentation=tiles_presentation, status_metrics=status_metrics
-    )
-    return tuple(label for label in picks if label)[:max_slots]
-
-
-def _sanitize_card_stat_picks(
-    picks: list[str],
-    *,
-    available: frozenset[str],
-    max_slots: int,
-) -> list[str]:
-    seen: set[str] = set()
-    out: list[str] = []
-    for raw in picks:
-        label = raw.strip() if isinstance(raw, str) else str(raw).strip()
-        if label and label in available and label not in seen:
-            seen.add(label)
-            out.append(label)
-        if len(out) >= max_slots:
-            break
-    return out
-
-
-def _ensure_card_stat_picks(
-    layout: LayoutId,
-    status_metrics: list[tuple[str, str]],
-    fmt: FormatId,
-    period_kind: PeriodKind,
-    *,
-    data_scope: str,
-    geo_scope: ShareSummaryGeoScope,
-    tiles_presentation: TilesPresentationId = "grid",
-) -> list[str]:
-    """Initialize or sanitize session picks for *layout*; returns UI row values."""
-    circle_cluster = _tiles_circle_cluster_picker(layout, tiles_presentation)
-    max_slots = _card_stat_max_slots(
-        layout, fmt, tiles_presentation=tiles_presentation, status_metrics=status_metrics
-    )
-    storage_max = tiles_circle_cluster_max(fmt) if circle_cluster else layout_card_stat_storage_max(layout)
-    available = frozenset(label for label, _ in status_metrics)
-    picks_key = _card_stat_picks_key(layout, period_kind)
-    count_key = _card_stat_slot_count_key(layout, period_kind)
-    scope_key = _card_stat_scope_key(layout, period_kind)
-    defaults = list(
-        default_card_stat_labels(
-            layout,
-            status_metrics,
-            period_kind=period_kind,
-            geo_scope=geo_scope,
-            fmt=fmt,
-        )
-    )
-
-    if st.session_state.get(scope_key) != data_scope:
-        st.session_state[scope_key] = data_scope
-        st.session_state.pop(picks_key, None)
-        st.session_state.pop(count_key, None)
-        _clear_card_stat_selectbox_keys(layout)
-
-    if picks_key not in st.session_state:
-        st.session_state[picks_key] = defaults
-        st.session_state[count_key] = _default_card_stat_slot_count(
-            circle_cluster=circle_cluster,
-            defaults=defaults,
-            max_slots=max_slots,
-            layout=layout,
-            fmt=fmt,
-            tiles_presentation=tiles_presentation,
-        )
-
-    sanitized = _sanitize_card_stat_picks(
-        list(st.session_state[picks_key]),
-        available=available,
-        max_slots=storage_max,
-    )
-    if not sanitized and defaults:
-        _clear_card_stat_selectbox_keys(layout)
-        st.session_state[picks_key] = defaults
-        st.session_state[count_key] = _default_card_stat_slot_count(
-            circle_cluster=circle_cluster,
-            defaults=defaults,
-            max_slots=max_slots,
-            layout=layout,
-            fmt=fmt,
-            tiles_presentation=tiles_presentation,
-        )
-        sanitized = list(defaults)
-
-    sanitized = _sanitize_card_stat_picks(
-        list(st.session_state[picks_key]),
-        available=available,
-        max_slots=storage_max,
-    )
-    slot_count = int(st.session_state.get(count_key, max(1, len(sanitized))))
-    ui_rows = _card_stat_ui_row_count(
-        layout,
-        fmt,
-        slot_count=slot_count,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-
-    raw = list(st.session_state[picks_key])
-    compact = [p for p in raw if p]
-    picks = (compact + [""] * ui_rows)[:ui_rows]
-    st.session_state[count_key] = ui_rows
-
-    st.session_state[picks_key] = picks
-    return picks
-
-
-def _status_metrics_lookup(status_metrics: list[tuple[str, str]]) -> dict[str, str]:
-    return dict(status_metrics)
-
-
-def _stats_on_card(picks: list[str]) -> set[str]:
-    return {label for label in picks if label}
-
-
-def _card_can_accept_stat(
-    layout: LayoutId,
-    fmt: FormatId,
-    picks: list[str],
-    slot_count: int,
-    *,
-    tiles_presentation: TilesPresentationId = "grid",
-    status_metrics: list[tuple[str, str]] | None = None,
-) -> bool:
-    max_slots = _card_stat_max_slots(
-        layout, fmt, tiles_presentation=tiles_presentation, status_metrics=status_metrics
-    )
-    ui_rows = _card_stat_ui_row_count(
-        layout,
-        fmt,
-        slot_count=slot_count,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-    active = (picks + [""] * ui_rows)[:ui_rows]
-    if any(not label for label in active):
-        return True
-    return ui_rows < max_slots
-
-
-def _add_stat_to_card(
-    label: str,
-    *,
-    layout: LayoutId,
-    period_kind: PeriodKind,
-    fmt: FormatId,
-    tiles_presentation: TilesPresentationId,
-    available_stat_count: int | None = None,
-) -> None:
-    """Fill the next empty card slot, or append a row when allowed."""
-    picks_key = _card_stat_picks_key(layout, period_kind)
-    count_key = _card_stat_slot_count_key(layout, period_kind)
-    max_slots = _card_stat_max_slots(
-        layout,
-        fmt,
-        tiles_presentation=tiles_presentation,
-        available_stat_count=available_stat_count,
-    )
-    picks = list(st.session_state.get(picks_key, []))
-    slot_count = int(st.session_state.get(count_key, 1))
-    ui_rows = _card_stat_ui_row_count(
-        layout,
-        fmt,
-        slot_count=slot_count,
-        tiles_presentation=tiles_presentation,
-        available_stat_count=available_stat_count,
-    )
-    active = (picks + [""] * ui_rows)[:ui_rows]
-
-    for i in range(ui_rows):
-        if not active[i]:
-            active[i] = label
-            st.session_state[picks_key] = active + picks[ui_rows:]
-            _clear_card_stat_selectbox_keys(layout)
-            return
-
-    if ui_rows < max_slots:
-        active.append(label)
-        st.session_state[count_key] = ui_rows + 1
-        st.session_state[picks_key] = active
-        _clear_card_stat_selectbox_keys(layout)
-
-
-def _add_stat_to_card_on_click(
-    stat_label: str,
-    layout: LayoutId,
-    period_kind: PeriodKind,
-    fmt: FormatId,
-    tiles_presentation: TilesPresentationId,
-    available_stat_count: int,
-) -> None:
-    """Callback — runs before widgets so selectbox keys can be cleared safely."""
-    _add_stat_to_card(
-        stat_label,
-        layout=layout,
-        period_kind=period_kind,
-        fmt=fmt,
-        tiles_presentation=tiles_presentation,
-        available_stat_count=available_stat_count,
-    )
-
-
-_CHIP_STRIP_COLS_PER_ROW = 3
-
-
-def _not_on_card_chip_strip(
-    *,
-    layout: LayoutId,
-    period_kind: PeriodKind,
-    fmt: FormatId,
-    status_metrics: list[tuple[str, str]],
-    picks: list[str],
-    slot_count: int,
-    metrics_lookup: dict[str, str],
-    tiles_presentation: TilesPresentationId = "grid",
-) -> None:
-    """Compact chips for stats not yet on the card; click to add."""
-    on_card = _stats_on_card(picks)
-    not_on_card = [
-        (stat_label, metrics_lookup[stat_label])
-        for stat_label, _ in status_metrics
-        if stat_label not in on_card
-    ]
-    if not not_on_card:
-        st.caption("All available stats are on the card.")
-        return
-
-    can_add = _card_can_accept_stat(
-        layout,
-        fmt,
-        picks,
-        slot_count,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-
-    available_stat_count = len(status_metrics)
-    for row_start in range(0, len(not_on_card), _CHIP_STRIP_COLS_PER_ROW):
-        row_items = not_on_card[row_start : row_start + _CHIP_STRIP_COLS_PER_ROW]
-        cols = st.columns(_CHIP_STRIP_COLS_PER_ROW)
-        for col_index in range(_CHIP_STRIP_COLS_PER_ROW):
-            with cols[col_index]:
-                if col_index >= len(row_items):
-                    continue
-                stat_label, value = row_items[col_index]
-                chip_index = row_start + col_index
-                st.button(
-                    f"{stat_label} · {value}",
-                    key=f"design_stat_chip_{layout}_{period_kind}_{chip_index}",
-                    type="tertiary",
-                    use_container_width=False,
-                    disabled=not can_add,
-                    on_click=_add_stat_to_card_on_click,
-                    args=(
-                        stat_label,
-                        layout,
-                        period_kind,
-                        fmt,
-                        tiles_presentation,
-                        available_stat_count,
-                    ),
-                )
-
-
-def _card_stat_picker_ui(
-    layout: LayoutId,
-    status_metrics: list[tuple[str, str]],
-    fmt: FormatId,
-    period_kind: PeriodKind,
-    *,
-    data_scope: str,
-    geo_scope: ShareSummaryGeoScope,
-    tiles_presentation: TilesPresentationId = "grid",
-) -> tuple[str, ...]:
-    """Ordered stat picker for tiles / list; hidden for spotlight."""
-    if layout in ("spotlight", "insight"):
-        return ()
-
-    max_slots = _card_stat_max_slots(
-        layout, fmt, tiles_presentation=tiles_presentation, status_metrics=status_metrics
-    )
-    circle_cluster = _tiles_circle_cluster_picker(layout, tiles_presentation)
-    available_labels = [label for label, _ in status_metrics]
-    metrics_lookup = _status_metrics_lookup(status_metrics)
-    if not available_labels:
-        st.caption("No statistics available for this period.")
-        return ()
-
-    picks_key = _card_stat_picks_key(layout, period_kind)
-    count_key = _card_stat_slot_count_key(layout, period_kind)
-    picks = _ensure_card_stat_picks(
-        layout,
-        status_metrics,
-        fmt,
-        period_kind,
-        data_scope=data_scope,
-        geo_scope=geo_scope,
-        tiles_presentation=tiles_presentation,
-    )
-    ui_rows = _card_stat_ui_row_count(
-        layout,
-        fmt,
-        slot_count=int(st.session_state[count_key]),
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-
-    if circle_cluster:
-        st.caption(
-            f"Supports up to {max_slots} stats. "
-            f"The default is {TILES_CIRCLE_CLUSTER_DEFAULT} circles."
-        )
-    elif layout == "tiles":
-        min_slots = _card_stat_min_slots(layout, fmt, tiles_presentation=tiles_presentation)
-        st.caption(
-            f"Statistics Grid: {min_slots}–{max_slots} stats "
-            f"({layout_grid_slot_limits_caption()}). "
-            "Use Add stat or the chips below to add more."
-        )
-    else:
-        st.caption(
-            f"Supports up to {max_slots} stats. "
-            "Use Add stat or the chips below to add more."
-        )
-
-    stat_row_cols = [0.5, 6, 1.8, 2.2]
-    min_slots = _card_stat_min_slots(layout, fmt, tiles_presentation=tiles_presentation)
-
-    for i in range(ui_rows):
-        current = picks[i] if i < len(picks) else ""
-        other = {picks[j] for j in range(len(picks)) if j != i and picks[j]}
-        options = [""] + [lab for lab in available_labels if lab not in other]
-        current = _sync_card_stat_selectbox_value(
-            layout, i, options=options, desired=current
-        )
-        picks[i] = current
-        can_up = i > 0
-        can_down = i < ui_rows - 1
-        can_remove = ui_rows > min_slots and (i > 0 or bool(current))
-
-        col_num, col_sel, col_val, col_actions = st.columns(
-            stat_row_cols,
-            vertical_alignment="center",
-        )
-        with col_num:
-            st.markdown(f"**{i + 1}**")
-        with col_sel:
-            choice = st.selectbox(
-                "Stat",
-                options=options,
-                format_func=lambda x: "—" if x == "" else x,
-                key=_card_stat_selectbox_key(layout, i),
-                label_visibility="collapsed",
-            )
-            picks[i] = choice or ""
-        with col_val:
-            value = metrics_lookup.get(picks[i], "") if picks[i] else ""
-            st.markdown(f"**{value}**" if value else "—")
-        with col_actions:
-            btn_up, btn_down, btn_rm = st.columns(3, gap="small")
-            with btn_up:
-                if st.button(
-                    "↑",
-                    key=f"design_card_stat_up_{layout}_{i}",
-                    help="Move up",
-                    disabled=not can_up,
-                    use_container_width=True,
-                ):
-                    picks[i - 1], picks[i] = picks[i], picks[i - 1]
-                    st.session_state[picks_key] = picks[:ui_rows]
-                    _clear_card_stat_selectbox_keys(layout)
-                    st.rerun()
-            with btn_down:
-                if st.button(
-                    "↓",
-                    key=f"design_card_stat_down_{layout}_{i}",
-                    help="Move down",
-                    disabled=not can_down,
-                    use_container_width=True,
-                ):
-                    picks[i + 1], picks[i] = picks[i], picks[i + 1]
-                    st.session_state[picks_key] = picks[:ui_rows]
-                    _clear_card_stat_selectbox_keys(layout)
-                    st.rerun()
-            with btn_rm:
-                if st.button(
-                    "✕",
-                    key=f"design_card_stat_rm_{layout}_{i}",
-                    help="Remove this stat",
-                    disabled=not can_remove,
-                    use_container_width=True,
-                ):
-                    if i == 0 and ui_rows == 1:
-                        picks[0] = ""
-                    elif i == 0:
-                        picks.pop(0)
-                        st.session_state[count_key] = ui_rows - 1
-                    else:
-                        picks.pop(i)
-                        st.session_state[count_key] = ui_rows - 1
-                    st.session_state[picks_key] = picks[:ui_rows]
-                    _clear_card_stat_selectbox_keys(layout)
-                    st.rerun()
-
-    col_num_foot, col_sel_foot, col_val_foot, col_actions_foot = st.columns(
-        stat_row_cols,
-        vertical_alignment="center",
-    )
-    with col_sel_foot:
-        if ui_rows < max_slots and st.button(
-            "Add stat",
-            key=f"design_card_stat_add_{layout}",
-        ):
-            st.session_state[count_key] = ui_rows + 1
-            st.rerun()
-    with col_actions_foot:
-        foot_up, foot_down, foot_rm = st.columns(3, gap="small")
-        with foot_down:
-            if st.button(
-                "Reset",
-                key=f"design_card_stat_reset_{layout}",
-                help="Restore this layout's default stat list",
-                use_container_width=True,
-            ):
-                defaults = list(
-                    default_card_stat_labels(
-                        layout,
-                        status_metrics,
-                        period_kind=period_kind,
-                        geo_scope=geo_scope,
-                        fmt=fmt,
-                    )
-                )
-                _clear_card_stat_selectbox_keys(layout)
-                st.session_state[picks_key] = defaults
-                st.session_state[count_key] = _default_card_stat_slot_count(
-                    circle_cluster=circle_cluster,
-                    defaults=defaults,
-                    max_slots=max_slots,
-                    layout=layout,
-                    fmt=fmt,
-                    tiles_presentation=tiles_presentation,
-                )
-                st.rerun()
-
-    slot_count = int(st.session_state[count_key])
-    can_add_more = _card_can_accept_stat(
-        layout,
-        fmt,
-        picks[:ui_rows],
-        slot_count,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-    stats_not_on_card = [
-        label for label in available_labels if label not in _stats_on_card(picks[:ui_rows])
-    ]
-    if stats_not_on_card and not can_add_more:
-        st.info("Card is full.")
-
-    st.session_state[picks_key] = picks[:ui_rows]
-    final = _effective_card_stat_labels(
-        picks,
-        layout,
-        fmt,
-        tiles_presentation=tiles_presentation,
-        status_metrics=status_metrics,
-    )
-    if not final:
-        st.caption("Select at least one stat to show on the card.")
-
-    st.divider()
-    _not_on_card_chip_strip(
-        layout=layout,
-        period_kind=period_kind,
-        fmt=fmt,
-        status_metrics=status_metrics,
-        picks=picks[:ui_rows],
-        slot_count=slot_count,
-        metrics_lookup=metrics_lookup,
-        tiles_presentation=tiles_presentation,
-    )
-    return final
-
-
-def _spotlight_label_from_session(
-    status_metrics: list[tuple[str, str]],
-) -> str:
-    available = {label for label, _ in status_metrics}
-    raw = st.session_state.get(_SPOTLIGHT_LABEL_KEY)
-    if isinstance(raw, str) and raw.strip() in available:
-        return raw.strip()
-    if SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT in available:
-        return SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT
-    if status_metrics:
-        return status_metrics[0][0]
-    return SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT
-
-
-def _select_spotlight_stat(stat_label: str) -> None:
-    """Callback — runs before widgets so the selectbox key can be updated."""
-    st.session_state[_SPOTLIGHT_LABEL_KEY] = stat_label
-
-
-def _spotlight_alternate_chip_strip(
-    status_metrics: list[tuple[str, str]],
-    *,
-    current: str,
-    metrics_lookup: dict[str, str],
-) -> None:
-    """Quick-switch chips for stats not currently spotlighted."""
-    others = [
-        (label, metrics_lookup[label])
-        for label, _ in status_metrics
-        if label != current
-    ]
-    if not others:
-        return
-    count = len(others)
-    st.caption(
-        f"{count} other stat{'s' if count != 1 else ''} available — click to spotlight."
-    )
-    cols_per_row = 3
-    for row_start in range(0, len(others), cols_per_row):
-        row_items = others[row_start : row_start + cols_per_row]
-        cols = st.columns(len(row_items))
-        for col_index, (stat_label, value) in enumerate(row_items):
-            with cols[col_index]:
-                chip_index = row_start + col_index
-                st.button(
-                    f"{stat_label} · {value}",
-                    key=f"design_spotlight_chip_{chip_index}",
-                    use_container_width=True,
-                    help=f"Spotlight {stat_label}",
-                    on_click=_select_spotlight_stat,
-                    args=(stat_label,),
-                )
-
-
-def _spotlight_stat_picker(status_metrics: list[tuple[str, str]]) -> None:
-    labels = [label for label, _ in status_metrics]
-    if not labels:
-        st.caption("No statistics available for this period.")
-        return
-    metrics_lookup = _status_metrics_lookup(status_metrics)
-    current = _spotlight_label_from_session(status_metrics)
-    col_sel, col_val = st.columns([4, 1], vertical_alignment="bottom")
-    with col_sel:
-        st.selectbox(
-            "Spotlight stat",
-            options=labels,
-            index=labels.index(current) if current in labels else 0,
-            key=_SPOTLIGHT_LABEL_KEY,
-            help="Single highlighted stat for the Spotlight layout.",
-        )
-    selected = _spotlight_label_from_session(status_metrics)
-    with col_val:
-        st.markdown(f"**{metrics_lookup.get(selected, '—')}**")
-
-    st.divider()
-    _spotlight_alternate_chip_strip(
-        status_metrics,
-        current=selected,
-        metrics_lookup=metrics_lookup,
-    )
-
-
-def _insight_fact_options(
-    facts: list[ShareSummaryInsightFact],
-    *,
-    species_options: tuple[str, ...] = (),
-) -> list[tuple[InsightFactId, str]]:
-    auto_ids: tuple[InsightFactId, ...] = (
-        "most_common_checklist_species",
-        "most_individuals_species",
-        "biggest_checklist_count",
-    )
-    options: list[tuple[InsightFactId, str]] = []
-    for fact_id in auto_ids:
-        if any(f.fact_id == fact_id for f in facts):
-            options.append((fact_id, INSIGHT_FACT_PICKER_LABELS[fact_id]))
-    if species_options:
-        options.append(("species_individuals", INSIGHT_FACT_PICKER_LABELS["species_individuals"]))
-    return options
-
-
-def _insight_fact_id_from_session(options: list[tuple[InsightFactId, str]]) -> InsightFactId:
-    valid = {fact_id for fact_id, _ in options}
-    raw = st.session_state.get(_INSIGHT_FACT_KEY, SHARE_SUMMARY_INSIGHT_FACT_DEFAULT)
-    if raw in valid:
-        return raw  # type: ignore[return-value]
-    if SHARE_SUMMARY_INSIGHT_FACT_DEFAULT in valid:
-        return SHARE_SUMMARY_INSIGHT_FACT_DEFAULT
-    return options[0][0] if options else SHARE_SUMMARY_INSIGHT_FACT_DEFAULT
-
-
-def _insight_species_from_session(species_options: tuple[str, ...]) -> str:
-    raw = st.session_state.get(_INSIGHT_SPECIES_KEY)
-    if isinstance(raw, str) and raw in species_options:
-        return raw
-    return species_options[0] if species_options else ""
-
-
-def _insight_fact_picker_ui(
-    facts: list[ShareSummaryInsightFact],
-    species_options: tuple[str, ...],
-    *,
-    df: pd.DataFrame,
-    period,
-) -> ShareSummaryInsightFact | None:
-    """Insights fact picker; returns the resolved fact for preview/export."""
-    options = _insight_fact_options(facts, species_options=species_options)
-    if not options:
-        st.caption("No insights available for this period.")
-        return None
-
-    fact_ids = [fact_id for fact_id, _ in options]
-    current_id = _insight_fact_id_from_session(options)
-    st.selectbox(
-        "Insights",
-        options=fact_ids,
-        format_func=lambda fid: INSIGHT_FACT_PICKER_LABELS[fid],
-        index=fact_ids.index(current_id) if current_id in fact_ids else 0,
-        key=_INSIGHT_FACT_KEY,
-        help="Species- and checklist-derived highlights for Interesting Insights cards.",
-    )
-    selected_id = _insight_fact_id_from_session(options)
-
-    species_common = ""
-    if insight_fact_requires_species(selected_id):
-        if not species_options:
-            st.caption("No species in this period for a selected-species fact.")
-            return None
-        species_common = _insight_species_from_session(species_options)
-        st.selectbox(
-            "Species",
-            options=list(species_options),
-            index=list(species_options).index(species_common) if species_common else 0,
-            key=_INSIGHT_SPECIES_KEY,
-        )
-        species_common = _insight_species_from_session(species_options)
-        refreshed = compute_insight_facts(
-            df,
-            period,
-            species_common=species_common,
-        )
-        return resolve_insight_fact(refreshed, selected_id)
-
-    return resolve_insight_fact(facts, selected_id)
-
-
-def _preview_scale_caption(fmt: FormatId, scale: float) -> str | None:
-    """Hint when preview is at or near export pixel size."""
-    if scale < _PREVIEW_SCALE_FULL - 0.005:
-        return None
-    width, height = FORMAT_PIXELS[fmt]
-    return (
-        f"Export size ({width}×{height}px). "
-        "Scroll the page to see the full card — story format is tall."
-        if height > width
-        else f"Export size ({width}×{height}px) — matches the PNG export."
-    )
-
-
-def _sidebar_preview_scale_controls(fmt: FormatId) -> float:
-    """Preview scale slider from compact default up to export pixel size."""
-    scale = st.slider(
-        "Preview scale",
-        min_value=_PREVIEW_SCALE_DEFAULT,
-        max_value=_PREVIEW_SCALE_FULL,
-        value=_PREVIEW_SCALE_DEFAULT,
-        step=0.01,
-        help="Drag right for full export size (100%). Default 0.42 is a compact preview.",
-        key="design_preview_scale",
-    )
-    caption = _preview_scale_caption(fmt, scale)
-    if caption:
-        st.caption(caption)
-    return scale
-
-
-def _centered_card_download_button(
-    *,
-    label: str,
-    data: bytes,
-    file_name: str,
-    mime: str,
-    help_text: str,
-) -> None:
-    """Download control centred under the scaled card preview."""
-    _, btn_col, _ = st.columns([1, 1, 1])
-    with btn_col:
-        st.download_button(
-            label,
-            data=data,
-            file_name=file_name,
-            mime=mime,
-            use_container_width=True,
-            help=help_text,
-        )
-
-
-@st.fragment
-def _current_card_fragment(
-    *,
-    stats: ShareSummaryStats,
-    all_time: ShareSummaryAllTimeStats | None,
-    selected_layout: LayoutId,
-    fmt: FormatId,
-    scale: float,
-    status_metrics: list[tuple[str, str]],
-    color_scheme_index: int,
-    card_stat_data_scope: str,
-    scope_label: str,
-    geo_scope: ShareSummaryGeoScope,
-    tiles_presentation: TilesPresentationId = "grid",
-    spotlight_presentation: SpotlightPresentationId = "classic",
-    insight_facts: list[ShareSummaryInsightFact],
-    insight_species_options: tuple[str, ...],
-    df_scoped: pd.DataFrame,
-    resolved_period,
-) -> None:
-    """Card statistics controls, live preview, and PNG export."""
-    card_stat_labels: tuple[str, ...] = ()
-    resolved_fact: ShareSummaryInsightFact | None = None
-    with st.expander(_STATISTICS_LABEL, expanded=False):
-        if selected_layout == "spotlight":
-            _spotlight_stat_picker(status_metrics)
-        elif selected_layout == "insight":
-            resolved_fact = _insight_fact_picker_ui(
-                insight_facts,
-                insight_species_options,
-                df=df_scoped,
-                period=resolved_period,
-            )
-        else:
-            card_stat_labels = _card_stat_picker_ui(
-                selected_layout,
-                status_metrics,
-                fmt,
-                stats.period_kind,
-                data_scope=card_stat_data_scope,
-                geo_scope=geo_scope,
-                tiles_presentation=tiles_presentation,
-            )
-
-    spotlight_label = _spotlight_label_from_session(status_metrics)
-
-    st.subheader(_CURRENT_CARD_LABEL)
-    st.markdown(
-        render_share_summary_preview_html(
-            stats,
-            layout=selected_layout,
-            fmt=fmt,
-            tiles_presentation=tiles_presentation,
-            spotlight_presentation=spotlight_presentation,
-            scale=scale,
-            spotlight_label=spotlight_label,
-            insight_fact=resolved_fact,
-            card_stat_labels=card_stat_labels,
-            all_time=all_time,
-            color_scheme_index=color_scheme_index,
-            geo_scope=geo_scope,
-            scope_label=scope_label,
-        ),
-        unsafe_allow_html=True,
-    )
-
-    png_filename = share_summary_png_filename(stats, layout=selected_layout, fmt=fmt)
-    try:
-        png_bytes = _cached_share_summary_png(
-            stats,
-            selected_layout,
-            fmt,
-            card_stat_labels,
-            spotlight_label,
-            all_time,
-            color_scheme_index,
-            share_summary_color_scheme_fingerprint(color_scheme_index),
-            scope_label,
-            geo_scope,
-            tiles_presentation,
-            spotlight_presentation,
-            resolved_fact,
-        )
-    except RuntimeError as exc:
-        st.warning(str(exc))
-    else:
-        _centered_card_download_button(
-            label="Export card",
-            data=png_bytes,
-            file_name=png_filename,
-            mime="image/png",
-            help_text="PNG of the current card above.",
-        )
-
-
 st.set_page_config(page_title=_DESIGN_STUDIO_TITLE, layout="wide")
 st.title(_DESIGN_STUDIO_TITLE)
 st.caption(
     "Exploratory tool only. Compare HTML mockups before choosing layouts for the main app."
 )
 
+keys = DESIGN_SOCIAL_CARDS_KEYS
 df: pd.DataFrame | None = None
 geo_scope = ShareSummaryGeoScope()
 
@@ -1325,49 +260,13 @@ with st.sidebar:
         )
 
     if df is not None and not df.empty:
-        geo_scope = _sidebar_geo_scope_controls(df)
+        geo_scope = render_sidebar_geo_scope_controls(df, keys)
     elif not use_sample and uploaded is None:
         st.caption("Upload a CSV to choose country and region.")
     elif df is not None and df.empty:
         st.caption("No data in file.")
 
-    st.header(_CURRENT_CARD_LABEL)
-    selected_layout: LayoutId = st.selectbox(
-        "Layout",
-        options=["tiles", "minimal", "spotlight", "insight"],
-        format_func=share_summary_layout_label,
-    )
-    tiles_presentation: TilesPresentationId = "grid"
-    if selected_layout == "tiles":
-        tiles_presentation = st.radio(
-            "Tiles presentation",
-            options=["grid", "circles"],
-            format_func=share_summary_tiles_presentation_label,
-            key=_TILES_PRESENTATION_KEY,
-            horizontal=True,
-        )
-    spotlight_presentation: SpotlightPresentationId = "classic"
-    if selected_layout == "spotlight":
-        spotlight_presentation = st.radio(
-            "Spotlight presentation",
-            options=["classic", "circle"],
-            format_func=share_summary_spotlight_presentation_label,
-            key=_SPOTLIGHT_PRESENTATION_KEY,
-            horizontal=True,
-        )
-    fmt: FormatId = st.selectbox(
-        "Aspect ratio",
-        options=["square", "portrait_post", "story"],
-        format_func=lambda x: FORMAT_LABELS[x],
-    )
-    color_theme_id = st.selectbox(
-        "Theme",
-        options=list(SHARE_SUMMARY_COLOR_SCHEME_IDS),
-        format_func=share_summary_color_scheme_label,
-        key=_COLOR_THEME_KEY,
-    )
-    color_scheme_index = share_summary_color_scheme_index(color_theme_id)
-    scale = _sidebar_preview_scale_controls(fmt)
+    sidebar_selection = render_sidebar_card_controls(keys)
 
 resolved_period = None
 stats: ShareSummaryStats = sample_share_summary_stats()
@@ -1472,31 +371,31 @@ if df is not None:
     )
 
 upload_name = uploaded.name if uploaded is not None else None
-card_stat_data_scope = _card_stat_data_scope(
+card_stat_data_scope = card_stat_data_scope_from_design_source(
     use_sample=use_sample,
     period_kind=stats.period_kind,
     period_label=stats.period_label,
     upload_name=upload_name,
     geo_scope=geo_scope,
-    fmt=fmt,
-    tiles_presentation=tiles_presentation,
+    fmt=sidebar_selection.fmt,
+    tiles_presentation=sidebar_selection.tiles_presentation,
 )
 scope_label = geo_scope_display_label(geo_scope)
 
 status_metrics = summary_status_metrics(stats, all_time=all_time, geo_scope=geo_scope)
 
-if _SPOTLIGHT_LABEL_KEY not in st.session_state:
-    st.session_state[_SPOTLIGHT_LABEL_KEY] = SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT
-if _INSIGHT_FACT_KEY not in st.session_state:
-    st.session_state[_INSIGHT_FACT_KEY] = SHARE_SUMMARY_INSIGHT_FACT_DEFAULT
+if keys.spotlight_label not in st.session_state:
+    st.session_state[keys.spotlight_label] = SHARE_SUMMARY_SPOTLIGHT_LABEL_DEFAULT
+if keys.insight_fact not in st.session_state:
+    st.session_state[keys.insight_fact] = SHARE_SUMMARY_INSIGHT_FACT_DEFAULT
 
 insight_facts: list[ShareSummaryInsightFact] = []
 insight_species_options: tuple[str, ...] = ()
 if df is not None and resolved_period is not None:
     insight_facts = compute_insight_facts(df_scoped, resolved_period)
     insight_species_options = species_common_names_in_period(df_scoped, resolved_period)
-    if insight_species_options and _INSIGHT_SPECIES_KEY not in st.session_state:
-        st.session_state[_INSIGHT_SPECIES_KEY] = insight_species_options[0]
+    if insight_species_options and keys.insight_species not in st.session_state:
+        st.session_state[keys.insight_species] = insight_species_options[0]
 
 tab_social_cards, tab_hex_experiments, tab_circle_layout = st.tabs(
     [
@@ -1512,7 +411,7 @@ hex_card_stat_labels = default_card_stat_labels(
     period_kind=stats.period_kind,
     geo_scope=geo_scope,
 )
-if fmt == "story" and len(hex_card_stat_labels) < SHARE_SUMMARY_STORY_MAX_STATS:
+if sidebar_selection.fmt == "story" and len(hex_card_stat_labels) < SHARE_SUMMARY_STORY_MAX_STATS:
     picked = set(hex_card_stat_labels)
     extra: list[str] = []
     for label, _ in status_metrics:
@@ -1522,35 +421,35 @@ if fmt == "story" and len(hex_card_stat_labels) < SHARE_SUMMARY_STORY_MAX_STATS:
         if len(hex_card_stat_labels) + len(extra) >= SHARE_SUMMARY_STORY_MAX_STATS:
             break
     hex_card_stat_labels = hex_card_stat_labels + tuple(extra)
-hex_preview_scale = min(scale, 0.32)
+hex_preview_scale = min(sidebar_selection.scale, 0.32)
 
 with tab_social_cards:
-    if not use_sample and not _period_has_checklist_data(stats):
+    if not use_sample and not period_has_checklist_data(stats):
         st.warning(
             f"No checklists in **{stats.period_label}** in this export. "
             "Try **Previous month** (or another range) in the sidebar, or pick a period that includes your data."
         )
 
-    _current_card_fragment(
+    render_current_card_fragment(
         stats=stats,
         all_time=all_time,
-        selected_layout=selected_layout,
-        fmt=fmt,
-        scale=scale,
+        selected_layout=sidebar_selection.layout,
+        fmt=sidebar_selection.fmt,
+        scale=sidebar_selection.scale,
         status_metrics=status_metrics,
-        color_scheme_index=color_scheme_index,
+        color_scheme_index=sidebar_selection.color_scheme_index,
         card_stat_data_scope=card_stat_data_scope,
         scope_label=scope_label,
         geo_scope=geo_scope,
-        tiles_presentation=tiles_presentation,
-        spotlight_presentation=spotlight_presentation,
+        keys=keys,
+        tiles_presentation=sidebar_selection.tiles_presentation,
+        spotlight_presentation=sidebar_selection.spotlight_presentation,
         insight_facts=insight_facts,
         insight_species_options=insight_species_options,
         df_scoped=df_scoped,
         resolved_period=resolved_period,
+        current_card_label=SOCIAL_CARDS_CURRENT_CARD_LABEL,
     )
-
-_DESIGN_HEX_SELECTED_KEY = "design_hex_selected_variant"
 
 with tab_circle_layout:
     st.caption(
@@ -1589,11 +488,11 @@ with tab_hex_experiments:
                 render_hex_grid_preview_html(
                     stats,
                     variant=variant,
-                    fmt=fmt,
+                    fmt=sidebar_selection.fmt,
                     scale=hex_preview_scale,
                     card_stat_labels=hex_card_stat_labels,
                     all_time=all_time,
-                    color_scheme_index=color_scheme_index,
+                    color_scheme_index=sidebar_selection.color_scheme_index,
                     geo_scope=geo_scope,
                     scope_label=scope_label,
                 ),
@@ -1604,17 +503,17 @@ with tab_hex_experiments:
     st.subheader("Selected preview (larger)")
     st.caption(
         f"{HEX_VARIANT_LABELS[selected_hex]} — use **Preview scale** in the sidebar "
-        f"(currently {scale:.0%} of export size)."
+        f"(currently {sidebar_selection.scale:.0%} of export size)."
     )
     st.markdown(
         render_hex_grid_preview_html(
             stats,
             variant=selected_hex,
-            fmt=fmt,
-            scale=scale,
+            fmt=sidebar_selection.fmt,
+            scale=sidebar_selection.scale,
             card_stat_labels=hex_card_stat_labels,
             all_time=all_time,
-            color_scheme_index=color_scheme_index,
+            color_scheme_index=sidebar_selection.color_scheme_index,
             geo_scope=geo_scope,
             scope_label=scope_label,
         ),
